HomeをSSRに変更
