+++ frontend/components/Header/index.tsx
@@ -4,7 +4,11 @@ import { JSX } from "react";
 import { Presenter } from "./Presenter";
 
 export const Header = async (): Promise<JSX.Element> => {
-  const cookieHeader = cookies().toString();
+  const cookieStore = await cookies();
+  const cookieHeader = cookieStore
+    .getAll()
+    .map((c) => `${c.name}=${c.value}`)
+    .join("; ");
   const user = await getMeFromRails(cookieHeader);
 
   return <Presenter user={user} />;
